backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
tree-sitter>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-java>=0.23.0
rank_bm25>=0.2.2
